move recursive_encode function to utils module.
